+++ cmd/kind/load/image-archive/image-archive.go
@@ -25,7 +25,7 @@ import (
 
 	"sigs.k8s.io/kind/pkg/cluster"
 	clusternodes "sigs.k8s.io/kind/pkg/cluster/nodes"
-	"sigs.k8s.io/kind/pkg/concurrent"
+	"sigs.k8s.io/kind/pkg/util/concurrent"
 )
 
 type flagpole struct {
