+++ pkg/internal/cluster/create/actions/config/config.go
@@ -26,11 +26,11 @@ import (
 
 	"sigs.k8s.io/kind/pkg/cluster/constants"
 	"sigs.k8s.io/kind/pkg/cluster/nodes"
-	"sigs.k8s.io/kind/pkg/concurrent"
 	"sigs.k8s.io/kind/pkg/internal/apis/config"
 	"sigs.k8s.io/kind/pkg/internal/cluster/create/actions"
 	"sigs.k8s.io/kind/pkg/internal/cluster/kubeadm"
 	"sigs.k8s.io/kind/pkg/kustomize"
+	"sigs.k8s.io/kind/pkg/util/concurrent"
 )
 
 // Action implements action for creating the kubeadm config
