+++ pkg/internal/cluster/create/actions/kubeadmjoin/join.go
@@ -28,10 +28,10 @@ import (
 
 	"sigs.k8s.io/kind/pkg/cluster/constants"
 	"sigs.k8s.io/kind/pkg/cluster/nodes"
-	"sigs.k8s.io/kind/pkg/concurrent"
 	"sigs.k8s.io/kind/pkg/exec"
 	"sigs.k8s.io/kind/pkg/fs"
 	"sigs.k8s.io/kind/pkg/internal/cluster/create/actions"
+	"sigs.k8s.io/kind/pkg/util/concurrent"
 )
 
 // Action implements action for creating the kubeadm join
